@@ -1,13 +1,14 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM  from 'react-dom/client';
 import App from './components/app/App';
 import './style/style.scss';
 
 
-ReactDOM.render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>,
-  document.getElementById('root')
+ReactDOM
+    .createRoot(document.getElementById('root'))
+    .render(
+        <React.StrictMode>
+          <App />
+        </React.StrictMode>
 );
 
