@@ -1,41 +1,24 @@
-import { useState } from "react";
+import { BrowserRouter as Router, Route, Routes} from "react-router-dom";
 import AppHeader from "../appHeader/AppHeader";
-import RandomChar from "../randomChar/RandomChar";
-import CharList from "../charList/CharList";
-import CharInfo from "../charInfo/CharInfo";
-import ErrorBoundary from "../errorBoundary/ErrorBoundary";
-
-import decoration from '../../resources/img/vision.png';
-import ComicsList from "../comicsList/ComicsList";
+import {MainPage, ComicsPage, Page404, SingleComicPage} from '../pages'
 
 const App = () => {
-    const [selectedChar, setChar] = useState(null)
-
-    const onCharSelected = id => {
-        setChar(id)
-    }
 
     return (
-        <div className="app">
-            <AppHeader/>
-            <main>
-                    {/* <ErrorBoundary>
-                        <RandomChar/>
-                    </ErrorBoundary>
-                <div className="char__content">
-                    <ErrorBoundary>
-                        <CharList onCharSelected = {onCharSelected}/>
-                    </ErrorBoundary>
-                    <ErrorBoundary>
-                        <CharInfo charId = {selectedChar}/>
-                    </ErrorBoundary>
-                </div>
-                <img className="bg-decoration" src={decoration} alt="vision"/> */}
-                    <ComicsList/>
-               
-            </main>
-        </div>
-    )
+        <Router>
+            <div className="app">
+                    <AppHeader/>
+                <main>
+                    <Routes>
+                        <Route  path="/" element={ <MainPage/>} />
+                        <Route  path="/comics" element={<ComicsPage/>} /> 
+                        <Route  path="/comics/:comicId" element={<SingleComicPage/>} />     
+                        <Route  path = '*' element={<Page404/>}/>
+                    </Routes>
+                </main>
+            </div>
+        </Router>
+        )
 }
 
 export default App;
