@@ -1,4 +1,5 @@
 import { useState, useEffect } from 'react';
+import {Link} from 'react-router-dom';
 import useMarvelService from '../../service/MarvelService';
 import Spinner from '../spinner/Spinner';
 import ErrorMessage from '../errorMassage/errorMassage';
@@ -6,12 +7,12 @@ import './comicsList.scss';
 
 const ComicsList = () => {
 
-    const [charList, setCharList] = useState([]);
+    const [comicsList, setComicsList] = useState([]);
     const [newItemLoading, setNewItemLoading] = useState(false);
     const [offset, setOffset] = useState(210);
-    const [charEnded, setCharEnded] = useState(false);
+    const [comicsEnded, setComicsEnded] = useState(false);
 
-    const {loading, error, getComics} = useMarvelService();
+    const {loading, error, clearError, getAllComics} = useMarvelService();
 
     useEffect(() => {
         onRequest(offset, true)
@@ -19,33 +20,33 @@ const ComicsList = () => {
 
     const onRequest = (offset, initial) => {
         initial ? setNewItemLoading(false) : setNewItemLoading(true);
-        getComics(offset)
-            .then(onCharListLoaded);
+        getAllComics(offset)
+            .then(onComicsListLoaded);
     }
-    const onCharListLoaded = (newCharList) => {
+    const onComicsListLoaded = (newComicsList) => {
+        clearError();
         let ended = false;
-        if( newCharList.length < 8 ) {
+        if( newComicsList.length < 8 ) {
             ended = true;
         }
-        setCharList(charList => [...charList, ...newCharList]);
-        setNewItemLoading(newItemLoading => false);
-        setOffset(offset => offset + 8);
-        setCharEnded(charEnded => ended);
+        setComicsList([...comicsList, ...newComicsList]);
+        setNewItemLoading(false);
+        setOffset(offset + 8);
+        setComicsEnded(ended);
 
     }
 
     function renderItems(arr) {
-        const items =  arr.map((item,i) => {
-
+        const items =  arr.map((item) => {
             return (
             <li className="comics__item"
                 tabIndex={0}
-                key={i}>
-                <a href="#">
+                key={item.id}>
+                <Link to = {`/comics/${item.id}`}>
                     <img src={item.thumbnail} alt={item.title} className="comics__item-img"/>
                     <div className="comics__item-name">{item.title}</div>
                     <div className="comics__item-price">{item.prices > 0 ? item.prices+'$' : 'NOT AVAILABLE ' }</div>
-                </a>
+                </Link>
             </li>
             )
         });
@@ -56,10 +57,9 @@ const ComicsList = () => {
         )
     }
 
-        const items = renderItems(charList);
+        const items = renderItems(comicsList);
         const errorMessage = error ? <ErrorMessage/> : null;
         const spinner = loading && !newItemLoading ? <Spinner/> : null;
-
         return (
             <div className="comics__list">
                 {errorMessage}
@@ -69,7 +69,7 @@ const ComicsList = () => {
                         className="button button__main button__long"
                         onClick={()=> onRequest(offset)}
                         disabled={newItemLoading}
-                        style={{'display': charEnded ? 'none' :  'block'}}>
+                        style={{'display': comicsEnded  ? 'none' :  'block'}}>
                  <div className="inner">load more</div>
                     </button>
             </div>
